Function rcl_sys::rcl_get_publishers_info_by_topic

pub unsafe extern "C" fn rcl_get_publishers_info_by_topic(
    node: *const rcl_node_t, 
    allocator: *mut rcutils_allocator_t, 
    topic_name: *const c_char, 
    no_mangle: bool, 
    publishers_info: *mut rcl_topic_endpoint_info_array_t
) -> rcl_ret_t

Return a list of all publishers to a topic. **

The node parameter must point to a valid node.

The topic_name parameter must not be NULL.

When the no_mangle parameter is true, the provided topic_name should be a valid topic name

for the middleware (useful when combining ROS with native middleware (e.g. DDS) apps).

When the no_mangle parameter is false, the provided topic_name should follow

ROS topic name conventions.

In either case, the topic name should always be fully qualified.

Each element in the publishers_info array will contain the node name, node namespace,

topic type, gid and the qos profile of the publisher.

It is the responsibility of the caller to ensure that publishers_info parameter points

to a valid struct of type rcl_topic_endpoint_info_array_t.

The count field inside the struct must be set to 0 and the info_array field inside

the struct must be set to null.

\see rmw_get_zero_initialized_topic_endpoint_info_array

The allocator will be used to allocate memory to the info_array member

inside of publishers_info.

Moreover, every const char * member inside of

rmw_topic_endpoint_info_t will be assigned a copied value on allocated memory.

\see rmw_topic_endpoint_info_set_node_name and the likes.

However, it is the responsibility of the caller to

reclaim any allocated resources to publishers_info to avoid leaking memory.

\see rmw_topic_endpoint_info_array_fini

Attribute | Adherence

—————— | ———––

Allocates Memory | Yes

Thread-Safe | No

Uses Atomics | No

Lock-Free | Maybe [1]

[1] implementation may need to protect the data structure with a lock

\param[in] node the handle to the node being used to query the ROS graph

\param[in] allocator allocator to be used when allocating space for

       the array inside publishers_info

\param[in] topic_name the name of the topic in question

\param[in] no_mangle if true, topic_name needs to be a valid middleware topic name,

       otherwise it should be a valid ROS topic name

\param[out] publishers_info a struct representing a list of publisher information

\return RCL_RET_OK if the query was successful, or

\return RCL_RET_NODE_INVALID if the node is invalid, or

\return RCL_RET_INVALID_ARGUMENT if any arguments are invalid, or

\return RCL_RET_BAD_ALLOC if memory allocation fails, or

\return RCL_RET_ERROR if an unspecified error occurs. */