Function rcl_sys::rcl_guard_condition_init

pub unsafe extern "C" fn rcl_guard_condition_init(
    guard_condition: *mut rcl_guard_condition_t, 
    context: *mut rcl_context_t, 
    options: rcl_guard_condition_options_t
) -> rcl_ret_t

Initialize a rcl guard_condition. **

After calling this function on a rcl_guard_condition_t, it can be passed to

rcl_wait() and then concurrently it can be triggered to wake-up rcl_wait().

Expected usage:

#include <rcl/rcl.h>

// … error handling

rcl_guard_condition_t guard_condition = rcl_get_zero_initialized_guard_condition();

// … customize guard condition options

rcl_ret_t ret = rcl_guard_condition_init(

&guard_condition, context, rcl_guard_condition_get_default_options());

// … error handling, and on shutdown do deinitialization:

ret = rcl_guard_condition_fini(&guard_condition);

// … error handling for rcl_guard_condition_fini()

Attribute | Adherence

—————— | ———––

Allocates Memory | Yes

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[inout] guard_condition preallocated guard_condition structure

\param[in] context the context instance with which the guard condition

should be associated

\param[in] options the guard_condition’s options

\return RCL_RET_OK if guard_condition was initialized successfully, or

\return RCL_RET_ALREADY_INIT if the guard condition is already initialized, or

\return RCL_RET_NOT_INIT if the given context is invalid, or

\return RCL_RET_INVALID_ARGUMENT if any arguments are invalid, or

\return RCL_RET_BAD_ALLOC if allocating memory failed, or

\return RCL_RET_ERROR if an unspecified error occurs. */