Function rcl_sys::rcl_guard_condition_init_from_rmw

pub unsafe extern "C" fn rcl_guard_condition_init_from_rmw(
    guard_condition: *mut rcl_guard_condition_t, 
    rmw_guard_condition: *const rmw_guard_condition_t, 
    context: *mut rcl_context_t, 
    options: rcl_guard_condition_options_t
) -> rcl_ret_t

Same as rcl_guard_condition_init(), but reusing an existing rmw handle. **

In addition to the documentation for rcl_guard_condition_init(), the

rmw_guard_condition parameter must not be NULL and must point to a valid

rmw guard condition.

Also the life time of the rcl guard condition is tied to the life time of

the rmw guard condition.

So if the rmw guard condition is destroyed before the rcl guard condition,

the rcl guard condition becomes invalid.

Similarly if the resulting rcl guard condition is fini’ed before the rmw

guard condition, then the rmw guard condition is no longer valid.

Attribute | Adherence

—————— | ———––

Allocates Memory | Yes

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[inout] guard_condition preallocated guard_condition structure

\param[in] rmw_guard_condition existing rmw guard condition to reuse

\param[in] context the context instance with which the rmw guard condition

was initialized with, i.e. the rmw context inside rcl context needs to

match rmw context in rmw guard condition

\param[in] options the guard_condition’s options

\return #RCL_RET_OK if guard_condition was initialized successfully, or

\return #RCL_RET_ALREADY_INIT if the guard condition is already initialized, or

\return #RCL_RET_INVALID_ARGUMENT if any arguments are invalid, or

\return #RCL_RET_BAD_ALLOC if allocating memory failed, or

\return #RCL_RET_ERROR if an unspecified error occurs. */