Function rcl_sys::rcl_logging_rosout_fini_publisher_for_node

pub unsafe extern "C" fn rcl_logging_rosout_fini_publisher_for_node(
    node: *mut rcl_node_t
) -> rcl_ret_t

Deregisters a rosout publisher for a node and cleans up allocated resources **

Calling this for an rcl_node_t will destroy the rosout publisher on that node and remove it from

the logging system so that no more Log messages are published to this function.

Attribute | Adherence

—————— | ———––

Allocates Memory | Yes

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[in] node a valid rcl_node_t that the publisher will be created on

\return RCL_RET_OK if the logging publisher was finalized successfully, or

\return RCL_RET_NODE_INVALID if any arguments are invalid, or

\return RCL_RET_BAD_ALLOC if allocating memory failed, or

\return RCL_RET_ERROR if an unspecified error occurs. */