Function rcl_sys::rcl_logging_rosout_output_handler

pub unsafe extern "C" fn rcl_logging_rosout_output_handler(
    location: *const rcutils_log_location_t, 
    severity: c_int, 
    name: *const c_char, 
    timestamp: rcutils_time_point_value_t, 
    format: *const c_char, 
    args: *mut va_list
)

The output handler outputs log messages to rosout topics. **

When called with a logger name and log message this function will attempt to

find a rosout publisher correlated with the logger name and publish a Log

message out via that publisher. If there is no publisher directly correlated

with the logger then nothing will be done.

This function is meant to be registered with the logging functions for rcutils

Attribute | Adherence

—————— | ———––

Allocates Memory | Yes

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[in] location The pointer to the location struct or NULL

\param[in] severity The severity level

\param[in] name The name of the logger, must be null terminated c string

\param[in] timestamp The timestamp for when the log message was made

\param[in] format The list of arguments to insert into the formatted log message

\param[in] args argument for the string format */