Function rcl_sys::rcl_parse_arguments

pub unsafe extern "C" fn rcl_parse_arguments(
    argc: c_int, 
    argv: *const *const c_char, 
    allocator: rcl_allocator_t, 
    args_output: *mut rcl_arguments_t
) -> rcl_ret_t

Parse command line arguments into a structure usable by code. **

\sa rcl_get_zero_initialized_arguments()

ROS arguments are expected to be scoped by a leading --ros-args flag and a trailing double

dash token -- which may be elided if no non-ROS arguments follow after the last --ros-args.

Remap rule parsing is supported via -r/--remap flags e.g. --remap from:=to or -r from:=to.

Successfully parsed remap rules are stored in the order they were given in argv.

If given arguments {"__ns:=/foo", "__ns:=/bar"} then the namespace used by nodes in this

process will be /foo and not /bar.

\sa rcl_remap_topic_name()

\sa rcl_remap_service_name()

\sa rcl_remap_node_name()

\sa rcl_remap_node_namespace()

Parameter override rule parsing is supported via -p/--param flags e.g. --param name:=value

or -p name:=value.

The default log level will be parsed as --log-level level, where level is a name

representing one of the log levels in the RCUTILS_LOG_SEVERITY enum, e.g. info, debug,

warn, not case sensitive.

If multiple of these rules are found, the last one parsed will be used.

If an argument does not appear to be a valid ROS argument e.g. a -r/--remap flag followed by

anything but a valid remap rule, parsing will fail immediately.

If an argument does not appear to be a known ROS argument, then it is skipped and left unparsed.

\sa rcl_arguments_get_count_unparsed_ros()

\sa rcl_arguments_get_unparsed_ros()

All arguments found outside a --ros-args ... -- scope are skipped and left unparsed.

\sa rcl_arguments_get_count_unparsed()

\sa rcl_arguments_get_unparsed()

Attribute | Adherence

—————— | ———––

Allocates Memory | Yes

Thread-Safe | Yes

Uses Atomics | No

Lock-Free | Yes

\param[in] argc The number of arguments in argv.

\param[in] argv The values of the arguments.

\param[in] allocator A valid allocator.

\param[out] args_output A structure that will contain the result of parsing.

Must be zero initialized before use.

\return RCL_RET_OK if the arguments were parsed successfully, or

\return RCL_RET_INVALID_ROS_ARGS if an invalid ROS argument is found, or

\return RCL_RET_INVALID_ARGUMENT if any function arguments are invalid, or

\return RCL_RET_BAD_ALLOC if allocating memory failed, or

\return RCL_RET_ERROR if an unspecified error occurs. */