Function rcl_sys::rcl_publisher_is_valid

pub unsafe extern "C" fn rcl_publisher_is_valid(
    publisher: *const rcl_publisher_t
) -> bool

Return true if the publisher is valid, otherwise false. **

The bool returned is false if publisher is invalid.

The bool returned is true otherwise.

In the case where false is to be returned, an error message is set.

This function cannot fail.

Attribute | Adherence

—————— | ———––

Allocates Memory | No

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[in] publisher pointer to the rcl publisher

\return true if publisher is valid, otherwise false */