Function rcl_sys::rcl_return_loaned_message_from_publisher

pub unsafe extern "C" fn rcl_return_loaned_message_from_publisher(
    publisher: *const rcl_publisher_t, 
    loaned_message: *mut c_void
) -> rcl_ret_t

Return a loaned message previously borrowed from a publisher. **

The ownership of the passed in ros message will be transferred back to the middleware.

The middleware might deallocate and destroy the message so that the pointer is no longer

guaranteed to be valid after that call.

Attribute | Adherence

—————— | ———––

Allocates Memory | No

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[in] publisher Publisher to which the loaned message is associated.

\param[in] loaned_message Loaned message to be deallocated and destroyed.

\return RCL_RET_OK if successful, or

\return RCL_RET_INVALID_ARGUMENT if an argument is null, or

\return RCL_RET_UNIMPLEMENTED if the middleware does not support that feature, or

\return RCL_RET_PUBLISHER_INVALID if the publisher is invalid, or

\return RCL_RET_ERROR if an unexpected error occurs and no message can be initialized. */