Function rcl_sys::rcl_subscription_fini

pub unsafe extern "C" fn rcl_subscription_fini(
    subscription: *mut rcl_subscription_t, 
    node: *mut rcl_node_t
) -> rcl_ret_t

Finalize a rcl_subscription_t. **

After calling, the node will no longer be subscribed on this topic

(assuming this is the only subscription on this topic in this node).

After calling, calls to rcl_wait and rcl_take will fail when using this

subscription.

Additioanlly rcl_wait will be interrupted if currently blocking.

However, the given node handle is still valid.

Attribute | Adherence

—————— | ———––

Allocates Memory | Yes

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[inout] subscription handle to the subscription to be deinitialized

\param[in] node a valid (not finalized) handle to the node used to create the subscription

\return RCL_RET_OK if subscription was deinitialized successfully, or

\return RCL_RET_INVALID_ARGUMENT if any arguments are invalid, or

\return RCL_RET_SUBSCRIPTION_INVALID if the subscription is invalid, or

\return RCL_RET_NODE_INVALID if the node is invalid, or

\return RCL_RET_ERROR if an unspecified error occurs. */