Function rcl_sys::rcl_take_loaned_message

pub unsafe extern "C" fn rcl_take_loaned_message(
    subscription: *const rcl_subscription_t, 
    loaned_message: *mut *mut c_void, 
    message_info: *mut rmw_message_info_t, 
    allocation: *mut rmw_subscription_allocation_t
) -> rcl_ret_t

Take a loaned message from a topic using a rcl subscription. **

Depending on the middleware, incoming messages can be loaned to the user’s callback

without further copying.

The implicit contract here is that the middleware owns the memory allocated for this message.

The user must not destroy the message, but rather has to return it with a call to

\sa rcl_return_loaned_message to the middleware.

Attribute | Adherence

—————— | ———––

Allocates Memory | No

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[in] subscription the handle to the subscription from which to take

\param[inout] loaned_message a pointer to the loaned messages.

\param[out] message_info rmw struct which contains meta-data for the message.

\param[in] allocation structure pointer used for memory preallocation (may be NULL)

\return RCL_RET_OK if the loaned message sequence was taken, or

\return RCL_RET_INVALID_ARGUMENT if any arguments are invalid, or

\return RCL_RET_SUBSCRIPTION_INVALID if the subscription is invalid, or

\return RCL_RET_BAD_ALLOC if allocating memory failed, or

\return RCL_RET_SUBSCRIPTION_TAKE_FAILED if take failed but no error

    occurred in the middleware, or

\return RCL_RET_UNIMPLEMENTED if the middleware does not support that feature, or

\return RCL_RET_ERROR if an unspecified error occurs. */