Function rcl_sys::rcl_timer_cancel

pub unsafe extern "C" fn rcl_timer_cancel(timer: *mut rcl_timer_t) -> rcl_ret_t

Cancel a timer. **

When a timer is canceled, rcl_timer_is_ready() will return false for that

timer, and rcl_timer_call() will fail with RCL_RET_TIMER_CANCELED.

A canceled timer can be reset with rcl_timer_reset(), and then used again.

Calling this function on an already canceled timer will succeed.

Attribute | Adherence

—————— | ———––

Allocates Memory | No

Thread-Safe | Yes

Uses Atomics | Yes

Lock-Free | Yes [1]

[1] if atomic_is_lock_free() returns true for atomic_int_least64_t

\param[inout] timer the timer to be canceled

\return RCL_RET_OK if the last call time was retrieved successfully, or

\return RCL_RET_INVALID_ARGUMENT if any arguments are invalid, or

\return RCL_RET_TIMER_INVALID if the timer is invalid, or

\return RCL_RET_ERROR an unspecified error occur. */