Function rcl_sys::rcutils_logging_get_logger_leveln

pub unsafe extern "C" fn rcutils_logging_get_logger_leveln(
    name: *const c_char, 
    name_length: usize
) -> c_int

Get the level for a logger and its name length. **

Identical to rcutils_logging_get_logger_level() but without

relying on the logger name to be a null terminated c string.

Attribute | Adherence

—————— | ———––

Allocates Memory | No, provided logging system is already initialized

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[in] name The name of the logger

\param[in] name_length Logger name length

\return The level of the logger if it has been set, or

\return RCUTILS_LOG_SEVERITY_UNSET if unset, or

\return g_rcutils_logging_default_logger_level for name_length of 0, or

\return -1 on invalid arguments, or

\return -1 if an error occurred */