Function rcl_sys::rcutils_string_map_get_capacity

pub unsafe extern "C" fn rcutils_string_map_get_capacity(
    string_map: *const rcutils_string_map_t, 
    capacity: *mut usize
) -> rcutils_ret_t

Get the current capacity of the string map. **

This function will return the internal capacity of the map, which is the

maximum number of key value pairs the map could hold.

The capacity can be set initially with rcutils_string_map_init() or

with rcutils_string_map_reserve().

The capacity does not indicate how many key value paris are stored in the

map, the rcutils_string_map_get_size() function can provide that.

\param[in] string_map rcutils_string_map_t to be queried

\param[out] capacity capacity of the string map

\return RCUTILS_RET_OK if successful, or

\return RCUTILS_RET_INVALID_ARGUMENT for invalid arguments, or

\return RCUTILS_RET_STRING_MAP_INVALID if the string map is invalid, or

\return RCUTILS_RET_ERROR if an unknown error occurs */