Function rcl_sys::rmw_borrow_loaned_message

pub unsafe extern "C" fn rmw_borrow_loaned_message(
    publisher: *const rmw_publisher_t, 
    type_support: *const rosidl_message_type_support_t, 
    ros_message: *mut *mut c_void
) -> rmw_ret_t

Borrow a loaned ROS message. **

This ROS message is owned by the middleware, that will keep it alive (i.e. in valid

memory space) until the caller publishes it using rmw_publish_loaned_message() or

returns it using rmw_return_loaned_message_from_publisher().

Attribute | Adherence

—————— | ———––

Allocates Memory | Maybe

Thread-Safe | Yes

Uses Atomics | Maybe [1]

Lock-Free | Maybe [1]

[1] implementation defined, check implementation documentation.

\par Runtime behavior

To borrow a ROS message is a synchronous operation.

It is also non-blocking, but it is not guaranteed to be lock-free.

Generally speaking, implementations may synchronize access to internal resources using

locks but are not allowed to wait for events with no guaranteed time bound (barring

the effects of starvation due to OS scheduling).

\par Memory allocation

It is implementation defined whether memory will be allocated on borrow or not.

Check the implementation documentation to learn about memory allocation

guarantees when using ROS message loaning support.

\par Thread-safety

Publishers are thread-safe objects, and so are all operations on them except for finalization.

Therefore, it is safe to borrow ROS messages from the same publisher concurrently.

\pre Given publisher must be a valid publisher, as returned by rmw_create_publisher().

\pre Given type_support must be a valid rosidl message type support, matching the

one registered with the publisher on creation.

\param[in] publisher Publisher to which the loaned ROS message will be associated.

\param[in] type_support Message type support of the loaned ROS message.

\param[out] ros_message Pointer to type erased ROS message loaned by the middleware.

\return RMW_RET_OK if successful, or

\return RMW_RET_BAD_ALLOC if memory allocation fails, or

\return RMW_RET_INVALID_ARGUMENT if publisher is NULL, or

\return RMW_RET_INVALID_ARGUMENT if type_support is NULL, or

\return RMW_RET_INVALID_ARGUMENT if ros_message is NULL, or

\return RMW_RET_INVALID_ARGUMENT if *ros_message is not NULL (to prevent leaks), or

\return RMW_RET_INCORRECT_RMW_IMPLEMENTATION if publisher implementation identifier

does not match this implementation, or

\return RMW_RET_UNSUPPORTED if the implementation does not support ROS message loaning, or

\return RMW_RET_ERROR if an unexpected error occured. */