Function rcl_sys::rmw_create_subscription

pub unsafe extern "C" fn rmw_create_subscription(
    node: *const rmw_node_t, 
    type_support: *const rosidl_message_type_support_t, 
    topic_name: *const c_char, 
    qos_policies: *const rmw_qos_profile_t, 
    subscription_options: *const rmw_subscription_options_t
) -> *mut rmw_subscription_t

Create a subscription and return a handle to that subscription. **

This function can fail, and therefore return NULL, if:

node is not a valid non-null handle for this rmw implementation,

as returned by `rmw_create_node()`

type_support is a not valid non-null message type support, as returned by

`ROSIDL_GET_MSG_TYPE_SUPPORT()`

topic_name is not a valid non-null topic name, according to

`rmw_validate_full_topic_name()` if ROS namespace conventions apply

qos_profile is not a fully specified non-null profile i.e. no UNKNOWN policies

subscription_options is not a valid non-null option set, such as the one

returned by `rmw_get_default_subscription_options()`

memory allocation fails during subscription creation

an unspecified error occurs

Attribute | Adherence

—————— | ———––

Allocates Memory | Yes

Thread-Safe | No

Uses Atomics | Maybe [1]

Lock-Free | Maybe [1]

[1] rmw implementation defined, check the implementation documentation

\param[in] node Handle to node with which to register this subscription

\param[in] type_support Type support for the messages to be subscribed to

\param[in] topic_name Name of the topic to subscribe to, often a fully qualified

topic name unless qos_profile is configured to avoid ROS namespace conventions

i.e. to create a native topic subscription

\param[in] qos_profile QoS policies for this subscription

\param[in] subscription_options Options for configuring this subscription

\return rmw subscription handle, or NULL if there was an error */