Function rcl_sys::rmw_get_serialization_format

pub unsafe extern "C" fn rmw_get_serialization_format() -> *const c_char

Get the unique serialization format for this middleware. **

Return the format in which binary data is serialized.

One middleware can only have one encoding.

In contrast to the implementation identifier, the serialization format can be equal between

multiple RMW implementations.

This means, that the same binary messages can be deserialized by RMW implementations with the

same format.

\sa rmw_serialize

\sa rmw_deserialize

\return serialization format */