Function rcl_sys::rmw_return_loaned_message_from_publisher

pub unsafe extern "C" fn rmw_return_loaned_message_from_publisher(
    publisher: *const rmw_publisher_t, 
    loaned_message: *mut c_void
) -> rmw_ret_t

Return a loaned message previously borrowed from a publisher. **

Tells the middleware that a borrowed ROS message is no longer needed by the caller.

Ownership of the ROS message is given back to the middleware.

If this function fails early due to a logical error, such as an invalid argument,

the loaned ROS message will be left unchanged.

Otherwise, ownership of the ROS message will be given back to the middleware.

It is up to the middleware what will be made of the returned ROS message.

It is undefined behavior to use a loaned ROS message after returning it.

Attribute | Adherence

—————— | ———––

Allocates Memory | No

Thread-Safe | Yes

Uses Atomics | Maybe [1]

Lock-Free | Maybe [1]

[1] implementation defined, check implementation documentation.

\par Runtime behavior

To return a ROS message is a synchronous operation.

It is also non-blocking, but it is not guaranteed to be lock-free.

Generally speaking, implementations may synchronize access to internal resources using

locks but are not allowed to wait for events with no guaranteed time bound (barring

the effects of starvation due to OS scheduling).

\par Thread-safety

Publishers are thread-safe objects, and so are all operations on them except for finalization.

Therefore, it is safe to return borrowed ROS messages to the same publisher concurrently.

However, since ownership of the loaned ROS message is given back to the middleware and

this transfer is not synchronized, it is not safe to return the same loaned ROS message

concurrently.

\pre Given publisher must be a valid publisher, as returned by rmw_create_publisher().

\pre Given loaned_message must have been previously borrowed from the same publisher

using rmw_borrow_loaned_message().

\param[in] publisher Publisher to which the loaned ROS message is associated.

\param[in] loaned_message Type erased loaned ROS message to be returned.

\return RMW_RET_OK if successful, or

\return RMW_RET_INVALID_ARGUMENT if publisher is NULL, or

\return RMW_RET_INVALID_ARGUMENT if loaned_message is NULL, or

\return RMW_RET_INCORRECT_RMW_IMPLEMENTATION if publisher implementation identifier

does not match this implementation, or

\return RMW_RET_UNSUPPORTED if the implementation does not support ROS message loaning, or

\return RMW_RET_ERROR if an unexpected error occurs and no message can be initialized. */