Function rcl_sys::rmw_send_response

pub unsafe extern "C" fn rmw_send_response(
    service: *const rmw_service_t, 
    request_header: *mut rmw_request_id_t, 
    ros_response: *mut c_void
) -> rmw_ret_t

Send a ROS service response. **

Send a ROS service response to the service client, with matching QoS policies,

from which the previously taken ROS service request was originally sent.

Attribute | Adherence

—————— | ———––

Allocates Memory | Maybe

Thread-Safe | Yes

Uses Atomics | Maybe [1]

Lock-Free | Maybe [1]

[1] implementation defined, check implementation documentation.

\par Runtime behavior

It is implementation defined whether sending a ROS service response is a

synchronous or asynchronous, and blocking or non-blocking, operation.

However, asynchronous implementations are not allowed to access the

given ROS service request after this function returns.

Check the implementation documentation to learn about request behavior.

\par Memory allocation

It is implementation defined whether memory will be allocated on send or not.

For instance, implementations that serialize ROS service responses may need to

perform additional memory allocations when dealing with unbounded (dynamically-sized)

fields.

\par Thread-safety

Service servers are thread-safe objects, and so are all operations on them except for

finalization.

Therefore, it is safe to send responses using the same service server concurrently.

However:

Access to the given ROS service request header is read-only but it is not synchronized.

Concurrent `request_header` reads are safe, but concurrent reads and writes are not.

Access to the given ROS service response is read-only but it is not synchronized.

Concurrent `ros_request` reads are safe, but concurrent reads and writes are not.

\pre Given service must be a valid service server, as returned by rmw_create_service().

\pre Given request_header must be the one previously taken along with the ROS service

request to which we reply.

\pre Given ros_response must be a valid service response, whose type matches the

service type support registered with the service on creation.

\param[in] client Service server to send a response with.

\param[in] request_header Service response header, same as the one taken

with the corresponding ROS service request.

\param[in] ros_response ROS service response to be sent.

\return RMW_RET_OK if successful, or

\return RMW_RET_BAD_ALLOC if memory allocation fails, or

\return RMW_RET_INVALID_ARGUMENT if service is NULL, or

\return RMW_RET_INVALID_ARGUMENT if request_header is NULL, or

\return RMW_RET_INVALID_ARGUMENT if ros_response is NULL, or

\return RMW_RET_INCORRECT_RMW_IMPLEMENTATION if the service

implementation identifier does not match this implementation, or

\return RMW_RET_ERROR if an unexpected error occurs. */