Function rcl_sys::rmw_wait

pub unsafe extern "C" fn rmw_wait(
    subscriptions: *mut rmw_subscriptions_t, 
    guard_conditions: *mut rmw_guard_conditions_t, 
    services: *mut rmw_services_t, 
    clients: *mut rmw_clients_t, 
    events: *mut rmw_events_t, 
    wait_set: *mut rmw_wait_set_t, 
    wait_timeout: *const rmw_time_t
) -> rmw_ret_t

Waits on sets of different entities and returns when one is ready. **

This function adds middleware-specific conditions to the wait set and waits

until one or more become ready, or until the timeout is reached.

\remark Elapsed time is measured against the system clock.

Timeout granularity is thus bound to that of the aforementioned clock and,

depending on the underlying implementation, to that of platform-specific

APIs to sleep and/or wait.

\remark

The amount of time this function actually waits may be either above or

below the specified timeout.

Arrays contain type-erased, middleware-specific conditions associated with

waitable entities, which this function casts and adds to the wait set.

NULL entries in arrays prior to wait are considered invalid.

When the wait is over, entries in each array that correspond to

conditions that were \b not triggered are set to NULL.

\remark Arrays’ memory management is external to this function.

Attribute | Adherence

—————— | ———––

Allocates Memory | Maybe [1]

Thread-Safe | No

Uses Atomics | Maybe [1]

Lock-Free | Maybe [1]

[1] rmw implementation defined, check the implementation documentation

\par Thread-safety

To wait is a reentrant procedure, but:

It is not safe to use the same wait set to wait in two or more threads concurrently.

It is not safe to wait for the same entity using different wait sets in two or

more threads concurrently.

Access to the given timeout is read-only but it is not synchronized.

Concurrent `wait_timeout` reads are safe, but concurrent reads and writes are not.

\pre Given wait_set must be a valid wait set, as returned by rmw_create_wait_set().

\pre All given entities must be associated with nodes that, in turn, were registered

with the same context the given wait_set was registered with on creation.

\param[inout] subscriptions Array of subscriptions to wait on.

Can be NULL if there are no subscriptions to wait on.

\param[inout] guard_conditions Array of guard conditions to wait on

Can be NULL if there are no guard conditions to wait on.

\param[inout] services Array of services to wait on.

Can be NULL if there are no services to wait on.

\param[inout] clients Array of clients to wait on.

Can be NULL if there are no clients to wait on.

\param[inout] events Array of events to wait on.

Can be NULL if there are no events to wait on.

\param[in] wait_set Wait set to use for waiting.

\param[in] wait_timeout If NULL, block indefinitely until an entity becomes ready.

If zero, do not block – check only for immediately available entities.

Else, this represents the maximum amount of time to wait for an entity to become ready.

\return RMW_RET_OK if successful, or

\return RMW_RET_TIMEOUT if wait timed out, or

\return RMW_RET_INVALID_ARGUMENT if wait_set is NULL, or

\return RMW_RET_INVALID_ARGUMENT if an array entry is NULL, or

\return RMW_RET_INCORRECT_RMW_IMPLEMENTATION if the wait_set implementation

identifier does not match this implementation, or

\return RMW_RET_ERROR if an unspecified error occurs. */