Module rcl_sys::rcl::graph

API in rcl/graph.h

Re-exports

pub use crate::rmw_get_zero_initialized_names_and_types as rcl_get_zero_initialized_names_and_types;

pub use crate::rmw_get_zero_initialized_topic_endpoint_info_array as rcl_get_zero_initialized_topic_endpoint_info_array;

pub use crate::rmw_names_and_types_t as rcl_names_and_types_t;

pub use crate::rmw_topic_endpoint_info_array_fini as rcl_topic_endpoint_info_array_fini;

pub use crate::rmw_topic_endpoint_info_array_t as rcl_topic_endpoint_info_array_t;

pub use crate::rmw_topic_endpoint_info_t as rcl_topic_endpoint_info_t;

Functions

rcl_count_publishers⚠

Return the number of publishers on a given topic.

rcl_count_subscribers⚠

Return the number of subscriptions on a given topic.

rcl_get_client_names_and_types_by_node⚠

Return a list of service client names and types associated with a node.

rcl_get_node_names⚠

Return a list of available nodes in the ROS graph.

rcl_get_node_names_with_enclaves⚠

Return a list of available nodes in the ROS graph, including their enclave names.

rcl_get_publisher_names_and_types_by_node⚠

Return a list of topic names and types for publishers associated with a node.

rcl_get_publishers_info_by_topic⚠

Return a list of all publishers to a topic.

rcl_get_service_names_and_types⚠

Return a list of service names and their types.

rcl_get_service_names_and_types_by_node⚠

Return a list of service names and types associated with a node.

rcl_get_subscriber_names_and_types_by_node⚠

Return a list of topic names and types for subscriptions associated with a node.

rcl_get_subscriptions_info_by_topic⚠

Return a list of all subscriptions to a topic.

rcl_get_topic_names_and_types⚠

Return a list of topic names and their types.

rcl_names_and_types_fini⚠

Finalize a rcl_names_and_types_t object.

rcl_names_and_types_init⚠

Initialize a rcl_names_and_types_t object.

rcl_service_server_is_available⚠

Check if a service server is available for the given service client.