Module rcl_sys::rcl::node

API for rcl/node.h

skip

rcl_node_get_rmw_handle

Structs

rcl_node_impl_t

rcl_node_t

Structure which encapsulates a ROS Node.

Functions

rcl_get_zero_initialized_node⚠

Return a rcl_node_t struct with members initialized to NULL.

rcl_node_fini⚠

Finalize a rcl_node_t.

rcl_node_get_domain_id⚠

Return the ROS domain ID that the node is using.

rcl_node_get_fully_qualified_name⚠

Return the fully qualified name of the node.

rcl_node_get_graph_guard_condition⚠

Return a guard condition which is triggered when the ROS graph changes.

rcl_node_get_logger_name⚠

Return the logger name of the node.

rcl_node_get_name⚠

Return the name of the node.

rcl_node_get_namespace⚠

Return the namespace of the node.

rcl_node_get_options⚠

Return the rcl node options.

rcl_node_get_rcl_instance_id⚠

Return the associated rcl instance id.

rcl_node_init⚠

Initialize a ROS node.

rcl_node_is_valid⚠

Return true if the node is valid, else false.

rcl_node_is_valid_except_context⚠

Return true if node is valid, except for the context being valid.