Module rcl_sys::rcl::subscription

API in rcl/subscription.rs

Structs

rcl_subscription_options_t

Options available for a rcl subscription.

rcl_subscription_t

Structure which encapsulates a ROS Subscription.

Functions

rcl_get_zero_initialized_subscription⚠

Return a rcl_subscription_t struct with members set to NULL.

rcl_return_loaned_message_from_subscription⚠

Return a loaned message from a topic using a rcl subscription.

rcl_subscription_can_loan_messages⚠

Check if subscription instance can loan messages.

rcl_subscription_fini⚠

Finalize a rcl_subscription_t.

rcl_subscription_get_actual_qos⚠

Get the actual qos settings of the subscription.

rcl_subscription_get_default_options⚠

Return the default subscription options in a rcl_subscription_options_t.

rcl_subscription_get_options⚠

Return the rcl subscription options.

rcl_subscription_get_publisher_count⚠

Get the number of publishers matched to a subscription.

rcl_subscription_get_topic_name⚠

Get the topic name for the subscription.

rcl_subscription_init⚠

Initialize a ROS subscription.

rcl_subscription_is_valid⚠

Check that the subscription is valid.“]

rcl_take⚠

Take a ROS message from a topic using a rcl subscription.

rcl_take_loaned_message⚠

“ Take a loaned message from a topic using a rcl subscription.

rcl_take_serialized_message⚠

Take a serialized raw message from a topic using a rcl subscription.