Module rcl_sys::rcl::types

API in rcl/types.h

Constants

RCL_RET_ALREADY_INIT

rcl_init() already called return code.

RCL_RET_ALREADY_SHUTDOWN

rcl_shutdown() already called return code.

RCL_RET_BAD_ALLOC

Failed to allocate memory return code.

RCL_RET_CLIENT_INVALID

Invalid rcl_client_t given return code.

RCL_RET_CLIENT_TAKE_FAILED

Failed to take a response from the client return code.

RCL_RET_ERROR

Unspecified error return code.

RCL_RET_EVENT_INVALID

Invalid rcl_event_t given return code.

RCL_RET_EVENT_TAKE_FAILED

Failed to take an event from the event handle

RCL_RET_INVALID_ARGUMENT

Invalid argument return code.

RCL_RET_INVALID_LOG_LEVEL_RULE

Argument is not a valid log level rule

RCL_RET_INVALID_PARAM_RULE

Argument is not a valid parameter rule

RCL_RET_INVALID_REMAP_RULE

Argument is not a valid remap rule

RCL_RET_INVALID_ROS_ARGS

Found invalid ros argument while parsing

RCL_RET_MISMATCHED_RMW_ID

Mismatched rmw identifier return code.

RCL_RET_NODE_INVALID

Invalid rcl_node_t given return code.

RCL_RET_NODE_INVALID_NAME

Invalid node name return code.

RCL_RET_NODE_INVALID_NAMESPACE

Invalid node namespace return code.

RCL_RET_NODE_NAME_NON_EXISTENT

Failed to find node name

RCL_RET_NOT_INIT

rcl_init() not yet called return code.

RCL_RET_OK

Success return code.

RCL_RET_PUBLISHER_INVALID

Invalid rcl_publisher_t given return code.

RCL_RET_SERVICE_INVALID

Invalid rcl_service_t given return code.

RCL_RET_SERVICE_NAME_INVALID

Service name (same as topic name) does not pass validation.

RCL_RET_SERVICE_TAKE_FAILED

Failed to take a request from the service return code.

RCL_RET_SUBSCRIPTION_INVALID

Invalid rcl_subscription_t given return code.

RCL_RET_SUBSCRIPTION_TAKE_FAILED

Failed to take a message from the subscription return code.

RCL_RET_TIMEOUT

Timeout occurred return code.

RCL_RET_TIMER_CANCELED

Given timer was canceled return code.

RCL_RET_TIMER_INVALID

Invalid rcl_timer_t given return code.

RCL_RET_TOPIC_NAME_INVALID

Topic name does not pass validation.

RCL_RET_UNKNOWN_SUBSTITUTION

Topic name substitution is unknown.

RCL_RET_UNSUPPORTED

Unsupported return code.

RCL_RET_WAIT_SET_EMPTY

Given rcl_wait_set_t is empty return code.

RCL_RET_WAIT_SET_FULL

Given rcl_wait_set_t is full return code.

RCL_RET_WAIT_SET_INVALID

Invalid rcl_wait_set_t given return code.

RCL_RET_WRONG_LEXEME

Expected one type of lexeme but got another

Type Definitions

rcl_ret_t

The type that holds an rcl return code.

rcl_serialized_message_t

typedef for rmw_serialized_message_t;