Module rcl_sys::rcutils::logging

API in rcutils/logging.h

Structs

rcutils_log_location_t

The structure identifying the caller location in the source code.

Enums

RcutilsLogSeverity

The severity levels of log messages / loggers.

Statics

g_rcutils_logging_initialized

The flag if the logging system has been initialized.

Functions

rcutils_log⚠

Log a message.

rcutils_logging_console_output_handler⚠

The default output handler outputs log messages to the standard streams.

rcutils_logging_get_default_logger_level⚠

Get the default level for loggers.

rcutils_logging_get_logger_effective_level⚠

Determine the effective level for a logger.

rcutils_logging_get_logger_level⚠

Get the severity level for a logger.

rcutils_logging_initialize⚠

Initialize the logging system.

rcutils_logging_initialize_with_allocator⚠

Initialize the logging system using the specified allocator.

rcutils_logging_logger_is_enabled_for⚠

Determine if a logger is enabled for a severity level.

rcutils_logging_set_default_logger_level⚠

Set the default severity level for loggers.

rcutils_logging_set_logger_level⚠

Set the severity level for a logger.

rcutils_logging_shutdown⚠

Shutdown the logging system.

Type Definitions

rcutils_logging_output_handler_t

The function signature to log messages.