Module rcl_sys::rmw::types

API in rmw/types.h

Structs

rmw_gid_t

ROS graph ID of the topic

rmw_liveliness_changed_status_t

QoS Liveliness Changed information provided by a subscription.

rmw_liveliness_lost_status_t

QoS Liveliness Lost information provided by a publisher.

rmw_message_info_t

Information describing an rmw message

rmw_offered_deadline_missed_status_t

QoS Deadline Missed information provided by a publisher.

rmw_publisher_allocation_t

Allocation of memory for an rmw publisher

rmw_publisher_options_t

Options that can be used to configure the creation of a publisher in rmw.

rmw_qos_profile_t

ROS MiddleWare quality of service profile.

rmw_request_id_t

An rmw service request identifier

rmw_requested_deadline_missed_status_t

QoS Requested Deadline Missed information provided by a subscription.

rmw_service_info_t

Meta-data for a service-related take.

rmw_subscription_allocation_t

Allocation of memory for an rmw subscription

rmw_subscription_options_t

Options that can be used to configure the creation of a subscription in rmw.

rmw_time_t

Struct representing a time point for rmw

Enums

RMWQoSDurabilityPolicy

QoS durability enumerations describing how samples persist

RMWQoSHistoryPolicy

QoS history enumerations describing how samples endure

RMWQoSLivelinessPolicy

QoS liveliness enumerations that describe a publisher’s reporting policy for its alive status. For a subscriber, these are its requirements for its topic’s publishers.

RMWQoSReliabilityPolicy

rmw_endpoint_type_t

Endpoint enumeration type

rmw_unique_network_flow_endpoints_requirement_t

Unique network flow endpoints requirement enumeration

Constants

RMWQoSDeadlineDefault

RMWQoSLifespanDefault

RMWQoSLivelinessLeaseDurationDefault

RMW_GID_STORAGE_SIZE

RMW_QOS_POLICY_DEPTH_SYSTEM_DEFAULT

Default size of the rmw queue when history is set to RMW_QOS_POLICY_HISTORY_KEEP_LAST, 0 indicates it is currently not set

Functions

rmw_get_zero_initialized_message_info⚠

Get zero initialized mesage info.

Type Definitions

rmw_time_point_value_t